Implement authorization

+++ apps/docs/app/api/comments/[id]/route.ts
@@ -1,6 +1,9 @@
 import { type NextRequest, NextResponse } from "next/server";
 import { z } from "zod";
+import { getServerSession } from "next-auth";
 import { db } from "@/utils/database";
+import { NOT_AUTHENTICATED } from "@/utils/errors";
+import { authOptions } from "../../auth/[...nextauth]/route";
 
 const patchSchema = z.strictObject({
   content: z.string().trim().min(1),
@@ -10,15 +13,22 @@ export async function PATCH(
   req: NextRequest,
   { params }: { params: { id: string } }
 ): Promise<NextResponse> {
+  const session = await getServerSession(authOptions);
+  if (!session?.user?.email) return NOT_AUTHENTICATED;
+
   const { content } = patchSchema.parse(await req.json());
 
   const target = await db
     .selectFrom("comments")
+    .select(["comments.author"])
     .where("id", "=", Number(params.id))
     .executeTakeFirst();
 
-  if (!target)
-    return NextResponse.json({ message: "Not Found" }, { status: 404 });
+  if (!target || target.author !== session.user.email)
+    return NextResponse.json(
+      { message: "Missing permissions" },
+      { status: 401 }
+    );
 
   await db
     .updateTable("comments")
@@ -33,13 +43,20 @@ export async function DELETE(
   _: NextRequest,
   { params }: { params: { id: string } }
 ): Promise<NextResponse> {
+  const session = await getServerSession(authOptions);
+  if (!session?.user?.email) return NOT_AUTHENTICATED;
+
   const target = await db
     .selectFrom("comments")
+    .select("author")
     .where("id", "=", Number(params.id))
     .executeTakeFirst();
 
-  if (!target)
-    return NextResponse.json({ message: "Not Found" }, { status: 404 });
+  if (!target || target.author !== session.user.email)
+    return NextResponse.json(
+      { message: "Missing permissions" },
+      { status: 401 }
+    );
 
   // TODO: Implement authentication
   await db.deleteFrom("comments").where("id", "=", Number(params.id)).execute();
